Extracted 'OnboardingTitle' View to it's own file
